// types.ts
export interface MarketNode { 
  id: string; 
  name: string; 
  lat: number; 
  lng: number; 
  cost: number; 
  desc: string; 
  assets: string[]; 
  region: 'america' | 'europe' | 'asia'; 
  timezone: string; 
}